Feature/optimizing coroutines (#89)
* Removed PlayerMover.MoveTo and PlayerMover.NavigateTo and made use of the db provided method directly.

* A lot of optimizations. Mainly for the TownRun and RiftCoroutine. There are still some issues marked with TODO.

* Setup expirience tracking.

* Greater Rift and Town Run should be working good in that commit. Cleaned up most of the Coroutines for the Townrun into a maintainable state.

* That stuff had no effect besides using up precious CPU Time...

* Some formating changes.

* Another readability change.

* More formatting changes. And TODO's placed in code.

* More formating changes.

* More formatting changes...

* Simplyfied CurrentRiftKeyCount.
Fixed the behavior of the TransmuteTag.

* More code formatting.

* More formatting changes...

* More formatting changes...

* Added a check for RiftCount.
Ported the cow level handling from the previous Couroutine.
The cow level needs a way to open it. I did not see it implemented previously.
The cow level needs to be reworked to use the quest you get instead of just looking for that npc.
Removed the old RiftCoroutine.

* Reenabled sanity checks in SceneStorage.

* Simplified UsePotion.

* Simplified VacuumItems.

* Added more checks in TrinityCombat.
Some codeformatting changes.

* Handling Gizmos higher up in the chain to prevent duplication of behavior.

* Started to implement the CoroutineResult convention.

* Cleaned up HookManager.

* Removed a ToList Enumeration during Stashing.

* Removed Obsolete Coroutine.

* Fixed a wrong condition.

* Formatting changes.

* Made UsePotion use CoroutineResult.

* Fixed TrinityGrid Update.
Removed a Wrapper around the TrinityGrid just to make it clearer what is going on.
Some code formatting.

* Code formatting changes.

* Some Messages that where listed in the Messages Tab fixed.

* Code formatting and reviews.

* Code formatting.

* Started to move stuff where it belongs.

* Removed the unused HandeStart method from IRoutine and RoutineBase.

* Cleaning out TrinityItem. Not gone completely yet.

* Handled an exception in AnimationBeamAvoidanceHandler.

* Missing file from previous commit.

* Changed the CombatHook to not use Trinity stuff but DB directly.

* Removed obsolete code related to wrapping Markers.

* Looting works again.

* Small optimization to LootProvider.
Added a missing file to the csproj.

* Fixed build errors in trinity.

* Fixed Looting.

* Rift: fixed the issue where the entry portal was detected as exit portal.

* Updated condition to match.

* Made sure to use the adjusted SNO stuff.

* Fixed a set of runtime errors.

* Rift: Made sure that there is no cycling between town and rift.

* Fixed transmute rare to legendary.

* Handling the python enum shizzle.

* Fixed a NullReferenceException in the Stuck Handler.

* Fixing an issue with Gizmo detection and Destructable handling.

* Some Code Cleanup.
Fixed some Floating point comparisons.
Removed some Logger references.
Prevented Stuck detection during stashing.

* Stashing treasure bags by default to loot them later on.

* Implemented a new profile tag to support bounties for now.

* Implemented SimpleBounties profile.

* Fixed a NullReferenceException.

* Code formatting and logging changes.

* Replaced log4net with Serilog -> better Performance.

* More logging.

* Made sure to unregister from an event when no longer required.

* Logging and code formatting.

* Increased log level of some important log messages.

* Disabled some logging for release builds.

* Changed the Definition that is used to enable TargetsCache logging.

* Made sure to reset previous visited level and world when rift is done.

* Fix for RiftKeys.

* Actually a Kamikaze fix but probably way more.

* The stop condition was implemented wrong and caused not enter portals.

* Handle stuck during Combat when it is standing still.

* Avoidance: No longer left and right stuck.

* Changed 2 log entries to provide more information.

* Using the new Log enricher to log the version.

* Changed the Plugin Metadata.

* Fixing some reviews.

+++ Components/Adventurer/Game/Quests/BountyCoroutineFactory.cs
@@ -36,7 +36,7 @@ namespace Trinity.Components.Adventurer.Game.Quests
                 return null;
             }
 
-            var bountyData = BountyDataFactory.GetBountyData((int)bountyInfo.Quest);
+            var bountyData = BountyDataFactory.GetBountyData(bountyInfo.Quest);
             if (bountyData == null)
             {
                 Core.Logger.Log("Unsupported bounty: {0} - {1}", (int)bountyInfo.Quest, bountyInfo.Info.DisplayName);
@@ -84,7 +84,7 @@ namespace Trinity.Components.Adventurer.Game.Quests
 
             if (bounty == null)
             {
-                bounty = new GenericBountyCoroutine((int)bountyInfo.Quest);
+                bounty = new GenericBountyCoroutine(bountyInfo.Quest);
                 //if (bounty.QuestData.Steps.First().Objectives.Any(o => o.ObjectiveType == QuestStepObjectiveType.EnterLevelArea))
                 //    return null;
                 //if (bounty.QuestData.QuestType != BountyQuestType.KillMonster && bounty.QuestData.QuestType != BountyQuestType.ClearZone)
